backend/requirements.txt
google-adk>=1.27.2
fastapi>=0.135.1
uvicorn>=0.42.0
sse-starlette>=3.3.3
python-dotenv>=1.2.2
pymupdf>=1.27.0
python-docx>=1.2.0
python-multipart>=0.0.22
